pyproject.toml
[build-system]
build-backend = 'hatchling.build'
requires = ['hatchling']

[project]
name = "typeitnow"
version = "0.1.2"
description = "A terminal-based typing game inspired by bop it."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "art>=6.5",
    "inputimeout>=1.0.4",
    "mido>=1.3.3",
    "pathlib>=1.0.1",
    "platformdirs>=4.5.0",
    "pyfluidsynth>=1.3.4",
]

[project.scripts]
typeitnow = "typeitnow.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/typeitnow"]
include = [
    "src/typeitnow/music/",
    "src/typeitnow/ascii/",
]

